Finish community and thread services

+++ code/services/thread-service/src/server.go
@@ -3,15 +3,18 @@ package server
 import (
 	"context"
 	"fmt"
+	communitypb "gen/community-service/pb"
 	dbpb "gen/db-service/pb"
 	threadpb "gen/thread-service/pb"
+
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
 type ThreadServer struct {
 	threadpb.UnimplementedThreadServiceServer
-	DBClient dbpb.DBServiceClient
+	CommunityClient communitypb.CommunityServiceClient
+	DBClient        dbpb.DBServiceClient
 }
 
 func (s *ThreadServer) ListThreads(ctx context.Context, req *threadpb.ListThreadsRequest) (*threadpb.ListThreadsResponse, error) {
@@ -56,6 +59,13 @@ func (s *ThreadServer) CreateThread(ctx context.Context, req *threadpb.CreateThr
 	userId, err := getCurrentUserId(ctx)
 	if err != nil {
 		return nil, err
+	}
+
+	_, err = s.CommunityClient.GetCommunity(ctx, &communitypb.GetCommunityRequest{
+		Id: req.CommunityId,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error calling community service: %w", err)
 	}
 
 	res, err := s.DBClient.CreateThread(ctx, &dbpb.CreateThreadRequest{
@@ -104,11 +114,20 @@ func (s *ThreadServer) UpdateThread(ctx context.Context, req *threadpb.UpdateThr
 		return nil, err
 	}
 
+	thread, err := s.DBClient.GetThread(ctx, &dbpb.GetThreadRequest{
+		Id: req.Id,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error calling database service: %w", err)
+	}
+	if thread.AuthorId != userId {
+		return nil, fmt.Errorf("user is not the author of the thread")
+	}
+
 	res, err := s.DBClient.UpdateThread(ctx, &dbpb.UpdateThreadRequest{
-		Id:       req.Id,
-		AuthorId: userId,
-		Title:    req.Title,
-		Content:  req.Content,
+		Id:      req.Id,
+		Title:   req.Title,
+		Content: req.Content,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error calling database service: %w", err)
@@ -131,9 +150,18 @@ func (s *ThreadServer) DeleteThread(ctx context.Context, req *threadpb.DeleteThr
 		return nil, err
 	}
 
+	thread, err := s.DBClient.GetThread(ctx, &dbpb.GetThreadRequest{
+		Id: req.Id,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error calling database service: %w", err)
+	}
+	if thread.AuthorId != userId {
+		return nil, fmt.Errorf("user is not the author of the thread")
+	}
+
 	_, err = s.DBClient.DeleteThread(ctx, &dbpb.DeleteThreadRequest{
-		Id:       req.Id,
-		AuthorId: userId,
+		Id: req.Id,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error calling database service: %w", err)
